+++ Flow/FluidGrid.h
@@ -16,7 +16,7 @@ namespace Flow
 	{
 	public:
 		FLOW_API FluidGrid();
-		FLOW_API FluidGrid(shared_ptr<ParticleVector> particles, int density, int sim_w, int sim_h);
+		FLOW_API FluidGrid(shared_ptr<ParticleVector> particles, int density, float sim_w, float sim_h);
 		FLOW_API ~FluidGrid();
 
 		FLOW_API void Update();
